src/requirements/dev.txt
-r base.txt
django-debug-toolbar==1.10.1